Update todo and refactor file locations
